+++ Screens/First-Aid Screens/Hypothermia.js
@@ -4,12 +4,11 @@ import { Container, Text, Content, Card, CardItem, Body, Left, View } from "nati
 import YoutubePlayer from 'react-native-youtube-iframe';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { StyleSheet } from "react-native";
-import Data from "../Data/Poisoning.json"
-export default function Poisoning() {
+import Data from "../../Data/Hypothermia.json"
+export default function Hypothermia() {
 
   const navigation = useNavigation();
-
-  function PoisoningNav() {
+    function HypothermiaNav() {
     navigation.navigate("Home");
   }
   var res = Data.filter(function(item) {
@@ -61,7 +60,7 @@ export default function Poisoning() {
             <YoutubePlayer
               height={300}
               play={false}
-              videoId={'b2ieb8BZJuY'}
+              videoId={'dEKaCOx7igI'}
             />
           </View>
         </Card>
@@ -88,3 +87,4 @@ const styles = StyleSheet.create({
     justifyContent: 'center'
   }
 });
+
